requirements.txt
weaviate-client
mysql
umap-learn
pandas
numpy
matplotlib
